chore(null): fixing all base use cases of null

+++ PocketKit/Sources/PocketKit/Home/HomeRecommendationCellViewModel.swift
@@ -56,7 +56,7 @@ extension HomeRecommendationCellViewModel: RecommendationCellViewModel {
 
     var timeToRead: String? {
         guard let timeToRead = recommendation.item?.timeToRead,
-              timeToRead > 0 else {
+              timeToRead.intValue > 0 else {
             return nil
         }
